สร้าง class ใหม่เสร็จเรียบร้อย
