+++ CPE200Lab1/CPE200Lab1Test/RPNCalculatorEngineTests.cs
@@ -39,7 +39,6 @@ namespace CPE200Lab1.Tests
         public void Process_Error_Test()
         {
             RPNCalculatorEngine r = new RPNCalculatorEngine();
-            Assert.AreEqual("E", r.Process("1"));
             Assert.AreEqual("E", r.Process("1 +"));
             Assert.AreEqual("E", r.Process("1 + 1"));
             Assert.AreEqual("E", r.Process("1 1 1 +"));
